add sample application with fabric smart client and token sdk (#19)
#### Type of change

- New feature

#### Description

This PR adds a sample application to the Fabric-X repo to make it easier
for developers to get started. In this first iteration it implements a
very basic fabric-smart-client plus fabric-token-sdk application.

It relies on the ansible scripts to deploy a local Fabric-X network. The
application can run in docker or natively, and is backwards compatible
with Fabric v3.

#### Related issues

N/A

---------

Signed-off-by: Arne Rutjes <arne123@gmail.com>
Co-authored-by: Marcus Brandenburger <bur@zurich.ibm.com>